feat(admin): split digital signage into its own Signage tab
Decouples the signage URL generator from the Super-Admin-only Export
tab into a standalone Signage tab, gated for any Admin. New order:
Diagnostics -> Signage -> Users -> Export.
