+++ frontend/styles/components/admin/SignageTab.module.scss
@@ -1,0 +1,111 @@
+@import '../../Variables.module.scss';
+
+.container {
+  display: flex;
+  flex-direction: column;
+  gap: 20px;
+  width: 100%;
+  box-sizing: border-box;
+}
+
+.cardDesc {
+  font-size: 14px;
+  color: #555;
+  margin-bottom: 14px;
+}
+
+.sectionLabel {
+  font-size: 16px;
+  font-weight: 600;
+  color: $medium-grey-color;
+  text-transform: uppercase;
+  letter-spacing: 0.02em;
+  margin: 20px 0 12px;
+
+  &:first-of-type {
+    margin-top: 0;
+  }
+}
+
+.filterGrid {
+  display: grid;
+  grid-template-columns: 1fr 1fr;
+  gap: 20px 40px;
+  margin-bottom: 20px;
+
+  @media (width <= 560px) {
+    grid-template-columns: 1fr;
+  }
+}
+
+.viewToggle {
+  display: inline-flex;
+  gap: 8px;
+  margin-bottom: 20px;
+}
+
+.viewButton {
+  background: #fff;
+  border: 1.5px solid #E5E5E5;
+  border-radius: 8px;
+  color: $black-color;
+  font-family: inherit;
+  font-size: 14px;
+  padding: 8px 18px;
+  cursor: pointer;
+}
+
+.viewButtonActive {
+  background: $brand-pink-color;
+  border-color: $brand-pink-color;
+  color: #fff;
+  font-weight: 600;
+}
+
+.linkRow {
+  display: flex;
+  gap: 12px;
+  margin-bottom: 8px;
+
+  @media (width <= $breakpoint-phone) {
+    flex-direction: column;
+  }
+}
+
+.linkField {
+  flex: 1;
+
+  // Without this, a flex item's default min-width: auto refuses to shrink below its content's
+  // intrinsic width -- a long signage URL would push .linkField (and the row) wider than the
+  // card instead of actually using the flex: 1 space it's been given.
+  min-width: 0;
+  overflow: hidden;
+  text-overflow: ellipsis;
+  white-space: nowrap;
+  font-family: monospace;
+  font-size: 13px;
+  color: $black-color;
+  background: $grey-lightest-color;
+  border: 1.5px solid #E5E5E5;
+  border-radius: 6px;
+  padding: 8px 12px;
+  box-sizing: border-box;
+}
+
+.copyButton {
+  background: $brand-pink-color;
+  color: #fff;
+  border: none;
+  border-radius: 8px;
+  padding: 8px 20px;
+  font-family: inherit;
+  font-size: 14px;
+  font-weight: 600;
+  cursor: pointer;
+  white-space: nowrap;
+}
+
+.linkCaption {
+  font-size: 13px;
+  color: $medium-grey-color;
+}
